lora-phy: Drop the timeout_in_ms argument for tx()
Firstly, this argument aws unused for sx127x radios, and on sx126x
radios it wasn't in milliseconds, instead each unit was 15.625us.

On sx126x timeout can be used as safety net to ensure that if Tx
is aborted or does not succeed (TxDone IRQ never triggers), radio
will trigger TxTimeout. But value given for the timeout should be
calculated for a given packet size, given modulation and packet
parameters.

Therefore in our case where we just pushed maximum value of 262s,
it was used wrongly anyway.