Cleanup JoinAcceptCreator, JoinRequestCreator and DataPayloadCreator. (#321)
* modify creators

This change makes the creator apis more uniform for JoinAcceptCreator, JoinRequestCreator and DataPayloadCreator.
The ::new api changes to accept a buffer instead of creating its own. This returns a result as the buffer might be too short.

This change also makes the new api a wrapper for with_options. lora-rs code uses the with_options directly as we want to pass in the generic CryptoFactory

JoinAcceptCreator new impl moved to default_crypto so it stays with the others.

* fix bug where the whole buffer was returned

* remove extra copy

* remove factory from creator struct

* use &DefaultFactory directly

* put examples back

* add tests for long buffers

* Update lorawan-encoding/tests/lorawan.rs

actually the test data must be checked as a slice so I think this needs to be left as is

Co-authored-by: Louis Thiery <thiery.louis@gmail.com>