encoding: Skip generating boilerplate for variable-length command creators
Macro-generated Creator implementation does not seem to be working
properly with variable-length commands.

As a workaround, stop generating Creator boilerplate for these and
provide a stubbed `UnimplementedCreator` trait as an interim solution.

This issue in current code only affects `McGroupStatusAns` from multicast
protocol, but there are also bunch of commands in certification protocol
which will be affected as well.