Method that allows for faster frequency updating while receiving.
Add [rx_switch_channel] method which performs frequency updating
conditional on the radio being in receive mode.

Previously, I believe this required the entire dance of creating
modulation params, creating rx packet params and preparing for rx
thereby resetting a lot of the same parameters over the SPI bus to
unchanged values.

This does the simplest thing of:
- entering standby mode
- updating just the frequency registers
- resuming rx

On an ESP32 + Sx1276 device, I observed this reducing frequency updating
times from 1815us to 490us on a release build (and 2710us to 1325us on a
debug build).